Add SimpleSessionProvider
Rename SimpleOvpSessionProvider to SimpleSessionProvider and recreate SimpleOvpSessionProvider as a subclass of SimpleSessionProvider.

SimpleOvpSessionProvider remains just to avoid breaking code that uses it.
